+++ resources/views/livro/show.blade.php
@@ -33,18 +33,18 @@
                 Ano: {{$livro->ano}}</p>
         </div>
         <div class="card-footer">
-            @auth
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
                 {{Form::open(['route' => ['livros.destroy',$livro->id],'method' => 'DELETE'])}}
                 @if ($nomeimagem !== "./img/livros/livrosemfoto.webp")
                 {{Form::hidden('foto',$nomeimagem)}}
                 @endif
                 <a href="{{url('livros/'.$livro->id.'/edit')}}" class="btn btn-success">Alterar</a>
                 {{Form::submit('Excluir',['class'=>'btn btn-danger','onclick'=>'return confirm("Confirma exclusão?")'])}}
-            @endauth
+            @endif
             <a href="{{url('livros/')}}" class="btn btn-secondary">Voltar</a>
-            @auth
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
                 {{Form::close()}}
-            @endauth
+            @endif
         </div>
     </div>
     <br />
